+++ src/plugins/flow/VirtualFlowActionHandler.ts
@@ -180,7 +180,8 @@ class DynamicFlowHandler extends ActionHandler {
 
         if (this.virtualDefaults) {
             if (this.virtualDefaults.mergeFunction) {
-                mergedOptions = this.virtualDefaults.mergeFunction(this.virtualDefaults.values, options);
+                const clonedDefaults = JSON.parse(JSON.stringify(this.virtualDefaults.values));
+                mergedOptions = this.virtualDefaults.mergeFunction(clonedDefaults, options);
             } else {
                 mergedOptions = DeepMergeUtil.merge(
                     this.virtualDefaults.values,
@@ -223,7 +224,6 @@ class DynamicFlowHandler extends ActionHandler {
         let metadata = FBLService.extractMetadata(this.action);
         metadata = await flowService.resolveOptionsWithNoHandlerCheck(context.ejsTemplateDelimiters.local, this.wd, metadata, context, false, parameters);
 
-        parameters = JSON.parse(JSON.stringify(parameters));
         parameters.parameters = this.getMergedOptions(options);
         parameters.wd = snapshot.wd;